requirements.txt
# V.A.R.G - Raspberry Pi Zero W Minimal Requirements
# Only keep packages available and lightweight on ARMv6

# NOTE: On Pi Zero W, prefer system packages via apt:
#   sudo apt install -y python3-numpy python3-pil python3-psutil python3-requests
# OpenCV should be installed via apt:
#   sudo apt install -y python3-opencv

numpy>=1.19.0,<1.26.0
Pillow>=8.0.0,<11.0.0
psutil>=5.8.0,<6.0.0
python-dotenv>=0.19.0,<2.0.0
requests>=2.25.0,<3.0.0
